+++ core/next.config.ts
@@ -24,11 +24,6 @@ const SettingsQuery = graphql(`
           cdnUrl
           checkoutUrl
         }
-        locales {
-          code
-          isDefault
-          path
-        }
       }
     }
   }
@@ -52,7 +47,6 @@ async function writeSettingsToBuildConfig() {
   }
 
   return await writeBuildConfig({
-    locales: data.site.settings?.locales,
     urls: {
       ...data.site.settings?.url,
       cdnUrls,